gui: scale down inner cube instead of moving faces out slight too far
This is beneficial because we use the same translation vectors to compute how to move tiles onto the correct part of their respective faces, which is technically very slightly wrong if those translation vectors are 1.001 instead of 1.0. Scaling down the inner cube marginally has the same effect of making sure the tiles appear in front of it, while not messing with the tiles position within their face. Very minor, but more correct!